fix: aggressive pre-roll render for stutter
As a workaround to resolve the stutter which is caused by delayed
Yospace ad event messaging and also schduling a new CSAI VPAID using
the Bitmovin API, a pre-emptive render aproach was adopted. This
starts to schedule the ad as soon as we receive the Yospace data,
but before ad events are actually surface which notify of that.

A couple of additional tweaks:
- Assume we only want to attempt to schedule the first VPAID in preroll
- Share some of the VPAID handling between aggressive and regular flow
- Breakout regular flow for clarity in a separate scheduler

Also added a feature flag for aggresive VPAIDs and to the demo
page to easily toggle between each of the scenarios.

Related to MECBM-659
Related to MECBM-592